pyproject.toml
[project]
name = "maventic-ai-101"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=1.5.0",
    "faiss-cpu>=1.13.2",
    "fastmcp>=2.14.5",
    "google-adk>=1.25.0",
    "html2text>=2025.4.15",
    "imbalanced-learn>=0.14.1",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.11",
    "langchain-groq>=1.1.2",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-ollama>=1.0.1",
    "litellm>=1.81.12",
    "lxml>=6.0.2",
    "matplotlib>=3.10.8",
    "mcp-server>=0.1.4",
    "mcp[cli]>=1.26.0",
    "numpy>=2.4.2",
    "ollama>=0.6.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "pyarrow>=23.0.0",
    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "xlrd>=2.0.2",
]
